image-service/requirements.txt
asgiref==3.3.1
click==7.1.2
Django==3.1.3
fastapi==0.61.2
h11==0.11.0
image==1.5.33
Pillow==8.0.1
pydantic==1.7.2
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
starlette==0.13.6
tk==0.1.0
uvicorn==0.12.3
